requirements.txt
flask==3.1.1
python-binance==1.0.22
python-dotenv==1.1.0
gunicorn==23.0.0
